Refactor: can react or unreact post
- Earlier, it was using independent react insert and delete by using
post_id. But, it's being done now by using post-react relationship.
- Also, reacts routes and tests have been moved to posts as it is always
connected to post
