+++ limbook_api/__init__.py
@@ -23,13 +23,11 @@ def create_app(config_class=Config):
     from limbook_api.main.routes import main
     from limbook_api.posts.routes import posts
     from limbook_api.comments.routes import comments
-    from limbook_api.reacts.routes import reacts
     from limbook_api.activities.routes import activities
     from limbook_api.image_manager.routes import image_manager
     app.register_blueprint(main)
     app.register_blueprint(posts)
     app.register_blueprint(comments)
-    app.register_blueprint(reacts)
     app.register_blueprint(activities)
     app.register_blueprint(image_manager)
